added plots and removed extra files
