+++ data/conll03/testa_beam1.csv
@@ -1,0 +1,64 @@
+ckpt,acc,prec,recall,f1
+ckpt-00-1.20359,82.10477335297995,19.711271515824542,5.978443920511957,9.174311926605505
+ckpt-04-0.78419,90.58125557408198,52.88072251635004,57.19097339171439,54.95145631067962
+ckpt-08-0.72048,94.10407538097638,73.52749755938822,76.1030650050522,74.79311486262827
+ckpt-12-0.73184,95.2634844313467,80.70356958096224,78.81441562815763,79.74780608332624
+ckpt-16-0.72240,95.71716623366552,82.52410468319559,80.71741327046144,81.61076111016516
+ckpt-20-0.69066,96.10492845786963,84.21143250688705,82.36780060626474,83.27941426868722
+ckpt-24-0.68400,96.35891271472333,85.23235800344234,83.3950825193668,84.30371127000342
+ckpt-32-0.65402,96.77575710574276,86.06668935011909,85.1970360390704,85.62965470548409
+ckpt-40-0.65225,96.8261661948893,85.51643584181545,86.30852138767263,85.9106529209622
+ckpt-48-0.65348,97.11311024080034,87.89590823489128,86.46008757157291,87.17208591561251
+ckpt-56-0.63622,97.24301058590872,88.06808510638298,87.13371505557427,87.59840853297214
+ckpt-64-0.65227,97.28954205281322,88.66758241758241,86.96530818457393,87.80819588505355
+ckpt-72-0.65337,97.32056303074954,88.48650495387768,87.23475917817447,87.85617367706921
+ckpt-80-0.63554,97.35158400868588,88.70249914412872,87.26844055237454,87.97962648556876
+ckpt-88-0.63593,97.37484974213812,88.64994026284349,87.47052879757494,88.05628549631261
+ckpt-96-0.66480,97.40974834231649,88.74127065235905,87.73997979117549,88.2377847404522
+ckpt-104-0.61710,97.40974834231649,88.95537698751923,87.62209498147524,88.28370238398234
+ckpt-112-0.64436,97.45821862034201,89.03665814151748,87.94206803637589,88.48597814115054
+ckpt-120-0.62483,97.40393190895342,88.76959781867757,87.72313910407544,88.24326613586312
+ckpt-128-0.64506,97.45821862034201,88.96845694799659,87.87470528797576,88.41819876302634
+ckpt-136-0.61922,97.39423785334833,88.4850543478261,87.73997979117549,88.11094199222053
+ckpt-144-0.60777,97.46015743146303,88.88132130086838,87.90838666217581,88.39217678435358
+ckpt-152-0.62663,97.47179029818915,88.98910823689586,88.05995284607611,88.52209243270696
+ckpt-160-0.63275,97.47372910931017,89.0287463854397,88.14415628157629,88.58424303968859
+ckpt-168-0.62913,97.45434099809997,88.95507147719537,88.02627147187604,88.48823429829015
+ckpt-176-0.64197,97.45434099809997,88.84360672440143,88.11047490737623,88.4755221104253
+ckpt-184-0.62059,97.44464694249486,88.79442271722496,87.94206803637589,88.36619003299772
+ckpt-192-0.60959,97.47372910931017,89.04972752043598,88.05995284607611,88.55207451312448
+ckpt-200-0.64437,97.45821862034201,88.97021276595744,88.02627147187604,88.49572504867518
+ckpt-208-0.63657,97.48730078715731,89.14042553191489,88.19467834287639,88.66503005163803
+ckpt-216-0.61485,97.48536197603629,89.05134308058483,88.21151902997643,88.62944162436548
+ckpt-224-0.62310,97.49893365388344,89.14596801633209,88.2452004041765,88.6932972241029
+ckpt-232-0.62094,97.50087246500446,89.07391673746814,88.27888177837656,88.67461727142012
+ckpt-240-0.63643,97.49311722052038,89.12155260469868,88.16099696867632,88.63867253640365
+ckpt-248-0.63298,97.52026057621467,89.20851063829788,88.26204109127652,88.73275205282316
+ckpt-256-0.62863,97.5047500872465,89.06648529161707,88.21151902997643,88.63694051950249
+ckpt-264-0.62986,97.48730078715731,89.1330267416113,88.12731559447626,88.62731814717588
+ckpt-272-0.62595,97.49311722052038,89.09121851599727,88.16099696867632,88.62366683595735
+ckpt-280-0.62223,97.4950560316414,89.18228279386712,88.16099696867632,88.66869918699187
+ckpt-288-0.62166,97.48342316491528,89.06462585034014,88.19467834287639,88.62751734642072
+ckpt-296-0.63228,97.4950560316414,89.0949302483838,88.19467834287639,88.64251861882194
+ckpt-304-0.62550,97.48536197603629,89.07977547201905,88.19467834287639,88.63501734788862
+ckpt-312-0.61433,97.5047500872465,89.16298060564817,88.26204109127652,88.71022342586325
+ckpt-320-0.63682,97.49893365388344,89.14596801633209,88.2452004041765,88.6932972241029
+ckpt-328-0.61981,97.4950560316414,89.15560095335377,88.19467834287639,88.6725364036573
+ckpt-336-0.59808,97.49311722052038,89.10605030591434,88.29572246547659,88.6990356961597
+ckpt-344-0.64506,97.48923959827833,89.0967851675455,88.21151902997643,88.65194211728866
+ckpt-352-0.62198,97.50281127612548,89.14412115024672,88.22835971707646,88.68387642826914
+ckpt-360-0.62728,97.49893365388344,89.15560095335377,88.19467834287639,88.6725364036573
+ckpt-368-0.62289,97.49699484276242,89.1670924885028,88.16099696867632,88.66119061732577
+ckpt-376-0.60663,97.49699484276242,89.15375446960667,88.17783765577636,88.66311065955466
+ckpt-384-0.61585,97.50862770948854,89.23522398228582,88.22835971707646,88.72893555762553
+ckpt-392-0.61403,97.50862770948854,89.17446808510638,88.22835971707646,88.69889105223059
+ckpt-400-0.62529,97.50862770948854,89.18964930200886,88.22835971707646,88.70640027091093
+ckpt-408-0.62610,97.51056652060956,89.14781425412485,88.26204109127652,88.7027164254887
+ckpt-416-0.61625,97.4950560316414,89.12340425531914,88.17783765577636,88.64809955134174
+ckpt-424-0.60771,97.4950560316414,89.13857677902621,88.17783765577636,88.65560447003048
+ckpt-432-0.61890,97.5047500872465,89.18780861569896,88.21151902997643,88.6969773939548
+ckpt-440-0.62231,97.50281127612548,89.15375446960667,88.17783765577636,88.66311065955466
+ckpt-448-0.62271,97.50862770948854,89.18780861569896,88.21151902997643,88.6969773939548
+ckpt-456-0.62908,97.49699484276242,89.13857677902621,88.17783765577636,88.65560447003048
+ckpt-464-0.60279,97.49311722052038,89.12155260469868,88.16099696867632,88.63867253640365
+ckpt-472-0.62532,97.49311722052038,89.12155260469868,88.16099696867632,88.63867253640365
