+++ requirements.txt
@@ -1,5 +1,6 @@
 absl-py==2.1.0
 asttokens==2.4.1
+attrs==24.2.0
 backcall==0.2.0
 cachetools==5.5.0
 certifi==2024.8.30
@@ -8,6 +9,7 @@ comm==0.2.2
 debugpy==1.8.5
 decorator==5.1.1
 executing==2.1.0
+fastjsonschema==2.20.0
 filelock==3.16.0
 fsspec==2024.9.0
 future==1.0.0
@@ -16,18 +18,23 @@ google-auth-oauthlib==1.0.0
 grpcio==1.66.1
 idna==3.8
 importlib_metadata==8.4.0
+importlib_resources==6.4.5
 ipykernel==6.29.5
 ipython==8.12.3
 ipywidgets==8.1.5
 jedi==0.19.1
 Jinja2==3.1.4
+jsonschema==4.23.0
+jsonschema-specifications==2023.12.1
 jupyter_client==8.6.2
 jupyter_core==5.7.2
 jupyterlab_widgets==3.0.13
+kaleido==0.2.1
 Markdown==3.7
 MarkupSafe==2.1.5
 matplotlib-inline==0.1.7
 mpmath==1.3.0
+nbformat==5.10.4
 nest-asyncio==1.6.0
 networkx==3.1
 numpy==1.24.4
@@ -47,9 +54,13 @@ oauthlib==3.2.2
 packaging==24.1
 pandas==2.0.3
 parso==0.8.4
+patsy==0.5.6
 pexpect==4.9.0
 pickleshare==0.7.5
+pkgutil_resolve_name==1.3.10
 platformdirs==4.3.2
+plotly==5.24.1
+plotly-express==0.4.1
 prompt_toolkit==3.0.47
 protobuf==5.28.0
 psutil==6.0.0
@@ -63,12 +74,17 @@ pytorch-lightning==1.1.1
 pytz==2024.1
 PyYAML==6.0.2
 pyzmq==26.2.0
+referencing==0.35.1
 requests==2.32.3
 requests-oauthlib==2.0.0
+rpds-py==0.20.0
 rsa==4.9
+scipy==1.10.1
 six==1.16.0
 stack-data==0.6.3
+statsmodels==0.14.1
 sympy==1.13.2
+tenacity==9.0.0
 tensorboard==2.14.0
 tensorboard-data-server==0.7.2
 torch==2.4.1
@@ -80,6 +96,7 @@ typing_extensions==4.12.2
 tzdata==2024.1
 urllib3==2.2.2
 wcwidth==0.2.13
+webcolors==24.8.0
 Werkzeug==3.0.4
 widgetsnbextension==4.0.13
 zipp==3.20.1
